+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # CHANGELOG
 
+## v1.0.2
+### Added
+- withPreReleaseDelimiter($preReleaseDelimiter)
+- removePreReleaseDelimiter()
+
+### Deprecations
+- withDelimiter($preReleaseDelimiter)
+
+  _replaced by withPreReleaseDelimiter($preReleaseDelimiter)_
+
+
+- removeDelimiter()
+
+  _replaced by removePreReleaseDelimiter()_
+
+
+# v1.0.1
+### Fixed
+- Licence model (Apache-2.0)
+- Version in composer.json
+
+---
+
 ## v1.0.0
 
 ### Added